+++ src/crc64/slice16.rs
@@ -51,7 +51,7 @@ impl<'a> Digest<'a, Slice16<u64>> {
 
 impl<'a> Hasher for Digest<'a, Slice16<u64>> {
     fn finish(&self) -> u64 {
-        self.clone().finalize() as u64
+        self.clone().finalize()
     }
 
     fn write(&mut self, bytes: &[u8]) {
